+++ frontend/src/tests/lib/components/accounts/WalletContextTest.svelte
